// src/app/models/starship.ts
export class Starship {
    cargo_capacity: number = 0;
    cost_in_credits: number = 0;
    crew: number = 0;
    hyperdrive_rating: number = 0;
    length: number = 0;
    manufacturer: string = "";
    model: string = "";
    name: string = "";
    passengers: number = 0;
    starship_class: string = "";
}